fix(tests): clean clippy --all-targets + isolate parallel test flake (bugs D-0/D-1/C-1)
D-0 (hard build-blockers): event_multi_type 3.14->3.5 (approx_constant); gate the
nightly g6_for_each_chunk bench behind a `nightly` cargo feature + required-features
so stable `--all-targets` no longer hits E0554. D-1 (~35 cosmetic test/bench lints):
doc_lazy_continuation / overindented lists, unnecessary i32->i32 casts, type_complexity
allows, field_reassign_with_default, assertions_on_constants allows, unused Without
import, while_let_on_iterator. C-1 (parallel flake): into_system_exclusive_smoke shared
a module static COUNTER across two tests; isolated each test's counter (also hardened
the command_queue test statics). All test/hygiene — lib edits confined to #[cfg(test)].
cargo clippy --workspace --all-targets -- -D warnings now clean on stable; 809 tests
pass across 3 parallel runs (flake gone).

+++ crates/boyko_ecs/benches/phase12_5_spawn_batch.rs
@@ -125,7 +125,7 @@ fn spawn_batch_10k_3comp(c: &mut Criterion) {
                                 y: (base + i) as f32,
                                 z: 0.0,
                             },
-                            health: Health((base + i) as i32),
+                            health: Health(base + i),
                         }))
                         .expect("5000 ≤ MAX_BATCH_HINT");
                 }
